+++ src/messages/unknown.rs
@@ -5,8 +5,8 @@ use super::{MessageParser, MessageData};
 pub struct UnknownMessage {
     source: String,
     target: String,
-    time: usize,
-    id: usize,
+    time: i64,
+    id: i64,
 }
 
 impl MessageData for UnknownMessage {
@@ -18,11 +18,11 @@ impl MessageData for UnknownMessage {
         &self.target
     }
 
-    fn time(&self) -> usize {
+    fn time(&self) -> i64 {
         self.time
     }
 
-    fn id(&self) -> usize {
+    fn id(&self) -> i64 {
         self.id
     }
 }
@@ -40,8 +40,8 @@ impl MessageParser for UnknownMessage {
         UnknownMessage {
             source: source.string(),
             target: target.string(),
-            id: id.number() as usize,
-            time: time.number() as usize,
+            id: id.number() as i64,
+            time: time.number() as i64,
         }
     }
 }
